requirements.txt
selenium>=4.15.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv